+++ src/error.rs
@@ -147,9 +147,6 @@ pub enum ClientError {
     /// Raised when disconnected from the server
     #[error("disconnected from server")]
     DisconnectedFromServer,
-    /// Raised when an invalid channel to send messages to the network handler is used
-    #[error("invalid channel to send messages to the network handler")]
-    InvalidChannel,
     /// Raised when [`Client::into_exclusive`](crate::client::Client::into_exclusive)
     /// is called while another handle on the same connection is still alive, so
     /// the connection an [`ExclusiveClient`](crate::client::ExclusiveClient)
@@ -365,7 +362,6 @@ impl ClientError {
             | ClientError::CannotParseBytes
             | ClientError::CannotParseEnum
             | ClientError::DisconnectedFromServer
-            | ClientError::InvalidChannel
             | ClientError::NotExclusive
             | ClientError::AlreadySubscribed
             | ClientError::UnexpectedSubscriptionConfirmation